Work in progress

+++ pico/cmsis-dsp-ex/process.h
@@ -8,7 +8,7 @@
 
 void do_fft();
 void process_init();
-void process(const uint16_t *inp, uint16_t *outp, uint16_t size);
+void process(const uint16_t *inp, uint8_t *outp, uint16_t size);
 void samples_to_float(const uint16_t *src, float *dest, uint16_t size, float scaling);
 
 #endif
